requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6
